+++ people.json
@@ -1,0 +1,1 @@
+"[#<Student:0x000001a162da75a0 @id=440, @age=\"23\", @name=\"Kelvin\", @parent_permission=true, @rentals=[]>]"
